moved payment api bit to client library
