--- src/main/java/uk/gov/hmcts/reform/ucmc/fnp/model/fee/FeeResponse.java
@@ -1,17 +1,0 @@
-package uk.gov.hmcts.reform.ucmc.fnp.model.fee;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Data;
-
-import java.math.BigDecimal;
-
-@Data
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class FeeResponse {
-    private String code;
-    private String description;
-    private Integer version;
-    @JsonProperty("fee_amount")
-    private BigDecimal amount;
-}
