--- src/main/java/uk/gov/hmcts/reform/ucmc/fnp/model/payment/enums/Currency.java
@@ -1,5 +1,0 @@
-package uk.gov.hmcts.reform.ucmc.fnp.model.payment.enums;
-
-public enum Currency {
-    GBP
-}
