--- src/main/java/uk/gov/hmcts/reform/ucmc/fnp/model/payment/enums/Service.java
@@ -1,5 +1,0 @@
-package uk.gov.hmcts.reform.ucmc.fnp.model.payment.enums;
-
-public enum Service {
-    FPL, CMC, DIVORCE, PROBATE, FINREM, DIGITAL_BAR
-}
