+++ src/main/resources/application.yaml
@@ -24,6 +24,12 @@ fees:
     channel: default
     event: issue
 
+payments:
+  api:
+    # TODO: update once f&p onboarding is completed
+    site_id: AA00
+    service: CMC
+
 idam:
   s2s-auth:
     microservice: ucmc_service
@@ -44,7 +50,7 @@ fees:
   api:
     url: http://localhost:8765
 
-payment:
+payments:
   api:
     url: http://localhost:8765
 
@@ -68,7 +74,7 @@ fees:
   api:
     url: http://fake-fees-register-api
 
-payment:
+payments:
   api:
     url: http://fake-payment-register-api
